+++ packages/elasticsearch/src/example-types/metricGroups/fieldValuesGroupStats.js
@@ -6,7 +6,7 @@ let { groupStats } = require('./groupStatUtils')
 let getSortField = field => {
   if (field === 'count') return '_count'
   if (field === 'key') return '_key'
-  return `${field}.value`
+  return field
 }
 
 let drilldown = ({ field, drilldown }, schema) => ({
@@ -21,7 +21,7 @@ let buildGroupQuery = (node, children, groupingType, schema) => {
     size = 10,
     filter,
     // sortField can be key, count, or stat name - min, max, avg, sum as long as its in stats
-    sort: { field: sortField, order = 'desc' } = {}, // todo: support array sort for multi-level
+    sort: { field: sortField, direction = 'desc' } = {}, // todo: support array sort for multi-level
   } = node
   let field = getField(schema, groupField)
   let query = {
@@ -30,7 +30,7 @@ let buildGroupQuery = (node, children, groupingType, schema) => {
         terms: {
           field,
           size,
-          ...(sortField && { order: { [getSortField(sortField)]: order } }),
+          ...(sortField && { order: { [getSortField(sortField)]: direction } }),
         },
         ...children,
       },